// eatin-client/src/pages/homePage/entities/categories.enum.ts
export enum Category {
    Egg = 'Egg Recipes',
    Cheela = 'Cheela Recipes (Chilla)',
    Waffle = 'Waffle Recipes',
    Crepe = 'Crepe Recipes',
    Thepla = 'Thepla Recipes',
    Cake = 'Cake Recipes',
    HomemadeBread = 'Homemade Bread Recipes',
    Soup = 'Soup Recipes',
    IndianNonVeg = 'Indian Non Veg Recipes',
    IndianHomemadeMasalaPowderAndChutneyPowder = 'Indian Homemade Masala Powder & Chutney Powder Recipes',
    BabyAndToddler10MonthOnwards = 'Baby & Toddler 10 Month Onwards Recipes',
    ThaiRecipesMalaysianAndEastAsian = 'Thai Recipes,Malaysian & East Asian Recipes',
    ChocolateCoffeeAndTeaDrink = 'Chocolate, Coffee & Tea Drink Recipes',
    Chicken = 'Chicken Recipes',
    IndianPicklesAndAchar = 'Indian Pickles & Achar Recipes',
    Idli = 'Idli Recipes',
    ContinentalPartyAppetizerAndStarter = 'Continental Party Appetizer & Starter Recipes',
    Omelette = 'Omelette Recipes',
    ContinentalBreakfast = 'Continental Breakfast Recipes',
    Sabzi = 'Sabzi Recipes',
    Kanji = 'Kanji Recipes (Porridge)',
    SriLankan = 'Sri Lankan Recipes',
    Paniyaram = 'Paniyaram Recipes',
    Dal = 'Dal Recipes',
    BabyAndToddler4To6Month ='Baby & Toddler 4-6 Month Recipes',
    Dosa = 'Dosa Recipes',
    Sambar = 'Sambar Recipes',
    ScrambledEgg = 'Scrambled Egg Recipe',
    GranolaCereal = 'Granola & Cereal Recipes',
    Cocktail = 'Cocktail Recipes',
    Kuzhambu = 'Kuzhambu Recipes',
    Adai = 'Adai Recipes',
    Toast = 'Toast Recipes',
    IndianSalad = 'Indian Salad Recipes',
    Uttapam = 'Uttapam Recipes',
    AppamAndIdiyappam = 'Appam Recipes & Idiyappam Recipes',
    Dhokla = 'Dhokla Recipes',
    Kootu = 'Kootu Recipes',
    IndianCurry = 'Indian Curry Recipes',
    BasicAndAdvancedCookingTechniques = 'Basic & Advanced Cooking Techniques',
    SweetIndianMithaiOrIndianDessert = 'Sweet Recipes (Indian Mithai / Indian Dessert)',
    SmoothiesandJuices = 'Smoothies and Juices',
    BreakfastBowl = 'Breakfast Bowl Recipes',
    Pulissery = 'Pulissery Recipe (Mor Kuzhambu)',
    IndianChinese = 'Indian Chinese Recipes',
    Roti= 'Roti Recipes',
    Kofta = 'Kofta Recipes',
    Pongal = 'Pongal Recipes',
    Kozhukattai = 'Kozhukattai Recipes (Pindi)',
    Thoran = 'Thoran Recipes',
    Rasam = 'Rasam Recipes',
    HomemadeEnergyBar = 'Homemade Energy Bar Recipes',
    Paratha = 'Paratha Recipes',
    Naan = 'Naan Recipes',
    Pachadi = 'Pachadi Recipes',
    Dessert = 'Dessert Recipes',
    Pesarattu = 'Pesarattu Recipes',
    BabyAndToddler6To9Month = 'Baby & Toddler 6-9 Month Recipes',
    Khichdi = 'Khichdi Recipes',
    World = 'World Recipes',
    Mutton = 'Mutton Recipes',
    Kulcha = 'Kulcha Recipes',
    Fish = 'Fish Recipes',
    HomemadeJamsSauceAndPesto = 'Homemade Jams, Sauce & Pesto Recipes',
    Pizza = 'Pizza Recipes',
    HashBrown = 'Hash Brown Recipes',
    IndianDrink = 'Indian Drink Recipes',
    Puttu = 'Puttu Recipes',
    Poriyal = 'Poriyal Recipes (South Indian Sabzi)',
    Kurma = 'Kurma Recipes (Korma)',
    IndianChutney = 'Indian Chutney Recipes',
    Porridge = 'Porridge Recipes',
    Drink = 'Drink Recipes',
    IndianRaita = 'Indian Raita Recipes',
    Biryani = 'Biryani Recipes',
    CookieAndBiscuit = 'Cookie & Biscuit Recipes',
    IndianSnack = 'Indian Snack Recipes',
    VarietyRice = 'Variety Rice (Mixed Rice/ Flavoured Rice Recipes)',
    Salad = 'Salad Recipes',
    IndianSideDish = 'Indian Side Dish Recipes',
    Pulao = 'Pulao Recipes',
    Erissery = 'Erissery Recipes',
    FrankieRoll = 'Frankie Roll Recipes (Wrap Recipes)',
    Palya = 'Palya Recipes',
    Pancake = 'Pancake Recipes',
    Pasta = 'Pasta Recipes',
    Burger = 'Burger Recipes',
    Puri = 'Puri Recipes',
    Mexican = 'Mexican Recipes',
    Poha = 'Poha Recipes',
    Upma = 'Upma Recipes',
    Sandwich = 'Sandwich Recipes',
    Kadhi = 'Kadhi Recipes',
    Snack = 'Snack Recipes',
    Parfait = 'Parfait Recipes',
    Continental = 'Continental Food Recipes',
    Gojju = 'Gojju Recipes',
    Stew = 'Stew Recipes',
    All = 'All Recipes'
}