// eatin-client/src/pages/homePage/entities/cuisines.enum.ts
export enum Cuisine {
    Kongunadu = 'Kongunadu',
    Rajasthani = 'Rajasthani',
    European = 'European',
    Jharkhand = 'Jharkhand',
    Lucknowi = 'Lucknowi',
    Awadhi = 'Awadhi',
    Indian = 'Indian',
    Jewish = 'Jewish',
    Kashmiri = 'Kashmiri',
    French = 'French',
    Sindhi = 'Sindhi',
    Asian = 'Asian',
    Thai = 'Thai',
    Karnataka = 'Karnataka',
    Singapore = 'Singapore',
    Burmese = 'Burmese',
    MiddleEastern = 'Middle Eastern',
    British = 'British',
    Korean = 'Korean',
    Maharashtrian = 'Maharashtrian Recipes',
    Punjabi = 'Punjabi',
    Sichuan = 'Sichuan',
    Gujarati = 'Gujarati Recipes﻿',
    Coorg = 'Coorg',
    NorthIndian = 'North Indian Recipes',
    Konkan = 'Konkan',
    Chinese = 'Chinese',
    CoastalKarnataka = 'Coastal Karnataka',
    Indonesian = 'Indonesian',
    Arab = 'Arab',
    Himachal = 'Himachal',
    Pakistani = 'Pakistani',
    Nepalese = 'Nepalese',
    Mughlai = 'Mughlai',
    Hyderabadi = 'Hyderabadi',
    Greek = 'Greek',
    Andhra = 'Andhra',
    Udupi = 'Udupi',
    SriLankan = 'Sri Lankan',
    Chettinad = 'Chettinad',
    Hunan = 'Hunan',
    American = 'American',
    Japanese = 'Japanese',
    Bengali = 'Bengali Recipes',
    UttarakhandNorthKumaon = 'Uttarakhand-North Kumaon ',
    NorthKarnataka = 'North Karnataka',
    Shandong = 'Shandong',
    Malabar = 'Malabar',
    Kerala = 'Kerala Recipes',
    Nagaland = 'Nagaland',
    NorthEastIndia = 'North East India Recipes',
    Mangalorean = 'Mangalorean',
    Haryana = 'Haryana',
    Bihari = 'Bihari',
    TamilNadu = 'Tamil Nadu',
    African = 'African',
    Swedish = 'Swedish',
    Goan = 'Goan Recipes',
    Cantonese = 'Cantonese',
    Vietnamese = 'Vietnamese',
    IndoChinese = 'Indo Chinese',
    Bangladeshi = 'Bangladeshi',
    Afghan = 'Afghan',
    Oriya = 'Oriya Recipes',
    Parsi = 'Parsi Recipes',
    Mexican = 'Mexican',
    SouthIndian = 'South Indian Recipes',
    Caribbean = 'Caribbean ',
    SouthKarnataka = 'South Karnataka',
    Assamese = 'Assamese',
    Malaysian = 'Malaysian',
    Fusion = 'Fusion',
    Mediterranean = 'Mediterranean',
    UttarPradesh = 'Uttar Pradesh',
    Continental = 'Continental',
    Italian = 'Italian Recipes',
    Malvani = 'Malvani'
}